Update supplier interface
